Initial scaffold of PowerVM virt driver
Initial driver that wraps the 'fake' implementation for PowerVM.

Also updates the tox.ini to match more of the pep8 rules that Nova
utilizes.

Provides an initial test case to validate that the driver can be
initialized.  This requires updates to the test framework to pull
in further dependencies of Nova.

Change-Id: I59407d14531635be567d2eea5d7e26ea895b5093