Integrate destroy baseline into driver
Provides a baseline for the destroy method in nova.  Updates the unit
tests as well.

Utilizes TaskFlow (but no roll back yet).  Future updates will provide
that capability.

Also updates the blockdev as there are new methods needed.  Had to
remove a rollback condition in the spawn due to these new methods.

Changed the way that the media deletion is done.  This new method should
be more efficient and complete.

Change-Id: I3909bd58af9b9a984f94f42de0d539c6d8b45fb0