Provide independent disk size from image
The image should only specify the size of the image.  The backing disk
may be independent of this.  For instance a 5 GB image put on to a 20 GB
disk.

This change set allows for that independence in the specification.  Will
create the disk based on the overall requirements from the flavor.

Adds a change to the localdisk.py to support this.

Minor MTMS fix for the available nodes in the driver.

Change-Id: I16cd312acf2d7cf395c55f5cfae0bddaf991a6a9