Initial refactoring for Volume Driver
The volume driver is a new construct that will connect the I/O Server to
a backing Cinder Volume.  This is a bit different from a standard
localdisk that will be provided via Glance.

This change set refactors the Glance disk work into a disk directory.
The cinder volume driver is then defined in a new 'volume' directory.
This will contain the driver volumes, such as NPIV and vSCSI.

The volume driver is built off the LibvirtBaseVolumeDriver model, but
can't extend that directly as that does contain libvirt specific code.
Eventually, we may be able to provide a common volume driver super class
that all drivers can utilize.

Change-Id: I2aafe2a97da8b195b74d57e8e88cb72389f69ac2