Initial cinder volume connector integration
This change set provides the volume connector for the nova_powervm
driver.  It currently gathers all of the WWPNs off of all VIOS.

The current implementation requires a restart of the agent if there are
changes to the VIOS (ex. a new VIOS is installed).

This change set also updates the behaviors within the local devices (ex.
non-Cinder backed).  If there is a boot cinder volume specified, then
the local disk will no longer create a disk.  This should be in line
with standard driver support.

This change set will allow a provision of a VM with a cinder backed
volume.  However, that volume is only visible to the Host server.  Not
to the VM as the connection between the VM and the hdisk is not part of
this change.

Change-Id: I3e24d8eb7b1f470b0123f93364777ed76a27d772