Rename the blockdev to disk driver
The original code had a 'blockdev' which represented local disk types.
However, with volume support, the term 'block' became ambiguous.  This
change set modifies the blockdev to be named as a 'driver' under the
disk folder.

This provides consistency with the volume driver.

Also updated the generic 'StorageAdapter' within blockdev to be a
'DiskAdapter'.  That more accurately describes its purpose.

Change-Id: I96f881109f6ad362c9fea2051bb38b79789d27f0