Provide Cinder rollback mechanism on taskflows
When an error occurs in the spawn or destroy tasks, the driver needs to
provide a rollback method to properly clean up the attachments.  This
change set provides such a mechanism.

This updates the cinder volume attach and detach to have rollbacks.
Also includes new unit tests to validate the rollback mechanism.

Change-Id: I2cf0d6520979c52759860d712d5286bc4ab19654