SSPDiskAdapter loads "the" SSP if no ssp_name
If no ssp_name is found in the config, SSPDiskAdapter will load the
entire SharedStoragePool feed from PowerVM.  If exactly one entry is
found in the feed, it will be used.

This change set also expands the custom exceptions used in the various
cases where we can't find an appropriate SSP to use.

Change-Id: I896c3ad9963a88764e7c89a4d212fbfce00e1a4f