NPIV Volume Connection
This change set provide a simple, single fabric, single FC port
connection for a given Virtual Machine.

This change puts the foundation in (via future change sets) for:
 - Delete connection
 - Multi port
 - Multi I/O Servers

This change does support booting from a NPIV volume.  It does not yet
support delete (though most of the code is in place).  This is due to
the volume connector not generating the same WWPNs.  That needs to be
stored in system metadata, which will be handled in a future change set.

Change-Id: If43f2d6afb56b6c6b6a2265ab0ce18f63d6f9eb4