Enhance NPIV connections for attach/detach volume
When attaching multiple volumes to an instance, the NPIV bus should
utilize the same NPIV connection multiple times.  It should do this to
ensure that there are not erroneous paths (and then virtual hardware) to
the VM.  This increases the performance to the VM as well.

This change set modifies how attach/detach work to fulfill that.

The system will now only create a VFC mapping if it does not already
exist.  It also will only delete the mappings upon deletion of the VM
itself (as the VM is what is serving as the host).

This also removes the vios components 'add_vfc_mapping'.  The reason for
this is, as we add complexity and flows (specifically with regard to
dual VIOS), a simple method will not provide the necessary operations.
And instead we need the flexibility in the npiv part to do quick and
other operations.

While this could have been put into the vios method, it would have
driven a lot of the 'npiv' logic down into that method.  Which did not
make sense.

This is also needed because as dual VIOS and multiple physical fabrics
come into play, we'll need to add more logic surrounding that.

Change-Id: I9536a708545fcbec24f7ecca7fe4034b62e20224