Initial multi-VIOS enablement
This is the first change set which gathers a set of Virtual I/O Server
elements.  None of the connection algorithms have been updated to make
use of multiple VIOSes yet.

This change set removes the vios_name from the conf options.  As most
operations will work against multiple VIOSes, and other conf options
will identify the VIOSes.

The new code generates a map of VIOS names to UUIDs.  It does not wrap
the whole objects because those become stale.

Change-Id: Id3c02d0f953181922f34a2758e41d11d141e5877