Migrate media to support multi VIOS
The media.py file is used to attach the config drive to the VM via a
virtual optical media.  Initially, the code required a single static
Virtual I/O Server that it could connect to.

However, PowerVM systems support multiple Virtual I/O Servers.  In case
one is offline, the code should simply move over to the other Virtual
I/O Server (if there is one or more available).

This change supports this scenario.  Changes because of this:
 - Removed the need for the vio_map to be passed into the media.py
 - Used static variables to significantly reduce the overhead of
   validation.
 - Consolidated the CreateCfgDrive and ConnectCfgDrive tasks into a
   single method.  This was required to maintain VIO precision upon the
   scsi mapping.

Change-Id: I39abf747b3437be15ce927b71eb9f1e53599609e