Cluster config for SSP disk driver
The SSP disk driver is going to need VIOS information in order to do
things like mapping LUs to LPARs.

The PowerVM SharedStoragePool object doesn't have Node (VIOS)
information, but the Cluster object does.

The Cluster object contains a reference to its corresponding
SharedStoragePool, but not vice versa.

So it makes more sense for the user to specify just the cluster name in
the configuration, and the SSP disk driver can discover the associated
SSP.

Change-Id: Ia7f88800eeeb1f576555fea777e2e4f8427ae808