Implement disk.ssp.create_disk_from_image
Implement create_disk_from_image method in the SSP disk driver.

o Move _get_disk_name to the disk.driver base class.
o Factor out algorithm to convert disk GB to bytes, with a floor for the
size of the image: disk.driver._disk_gb_to_bytes.
o Add method disk.ssp._any_vios_uuid to choose any VIOS from the
cluster.  This is to be used when we need to perform an operation
against a VIOS, but it doesn't matter which VIOS we choose.

Change-Id: I6a145ed7fbe6e78a76bea41e07c74d09c9c3e47a