create_disk_from_image with linked clone
Image LUs can't be mapped.  Disk LUs can't be uploaded.  In order to
create a disk for an LPAR, we have to make a Disk LU a linked clone to
an Image LU.

This change set augments create_disk_from_image to do the following:
o Check whether we already have an uploaded Image LU for this glance
image.  We key off of the LU name, which, for such Images, is
constructed with the keyword 'image' and the UUID of the glance image.
o If we don't already have the Image LU, create it and upload to it.
o Create a Disk LU linked clone to the Image LU.

Change-Id: I627d462fc06bc2307d39fde4bf85dc595551eab3