Implement get_host_ip_addr
One of the required driver methods is the get_host_ip_addr.  This should
return the IP Address where the driver is running.

The function of this method between PowerVM and KVM is nearly identical.
Therefore, this function was built off of the foundation from the KVM
implementation.  Citation included in the method text itself.

This change also includes corresponding unit tests.

Change-Id: Id193580d6a2860784ddf2ed5d9b99a03e96b776c