RMC NIC on each VM
RMC allows the management entity to provide directives to the virtual
machine, for things such as dynamically modifying resource allocations
(memory, CPU and I/O) as well as migration.  RMC runs over Ethernet.

This change set provides an additional NIC that will be deployed with
the VMs to provide the communication path for RMC.

Change-Id: I71aa328b021f0a19e27e5a7611df94896957e351