SSP Disk Driver: disconnect_image_disk
Implement disconnect_image_disk in the SSP disk driver.  This removes
the VSCSI mappings for all LUs (of the specified type) from all VIOSes
which are both a) on this host, and b) a member of the Cluster.

The return value is the list of LUs whose mappings were removed.  This
is used as input to the delete_disks method.  We must therefore take
into account that the list of mappings may indicate different sets of
LUs on each VIOS (if, for example, one VIOS was down when a particular
LU was added and mapped).

Verification will fail pending merge of the following from pypowervm:

pypowervm.tasks.scsi_mapper.remove_lu_mapping
Change-Id: Iab28fe072c949d08edb2d407b1e57a3a9f953173

Implement LU.__hash__
Change-Id: I8b04d8b60a427da6624f818cc4d38f27729b8cc2