Enhance VIOS logic to localdisk flow
Currently the VIOS logic in the localdisk would simply look for the
first Virtual I/O Server.

This enhancement enables the system to look through all of the VIOSes on
the system when determining the volume group to use.

An optional parameter was added to enable the user to specify exactly
which VIOS.  However, this is not required and is only needed if
multiple Virtual I/O Servers have a volume group of the same name.

Change-Id: I275384078507b85c1be884e05c08336e6cd5f534