Add storage XAG to npiv
The NPIV flow needs both the FC mapping data as well as the backing
storage devices.  Since pypowervm switched to a default of group=None,
that exposed an issue in the query in the npiv.py file where it was not
retrieving the Fibre Ports.

This update extends the XAG request to include Storage, as well as the
original FC_Mappings.

Change-Id: Iaf08b15183383fc1df8d9b294d2a18140e7b3e31