LPAR-specific SSP ops target VIOSes on same host
When nova is managing an SSP Cluster whose VIOSes span multiple
accessible hosts (e.g. nova is running against a single HMC managing
multiple hosts), it is important that operations specific to a
particular instance - such as creating and removing disk connections -
be invoked only on the VIOS(es) residing on the same host as that
instance.  This change set fixes the SSP Disk Driver to ensure this is
the case.

Ancillary to this change is an expansion and refactoring of the vm
utility to retrieve quick properties for an LPAR.  Invoking the new
vm.get_vm_qp method without the qprop parameter will return a dictionary
of all the quick properties available for the LPAR.

Change-Id: Ib1d02743baf041c9abc69b25201d33e516fa6cb3