Use the active WWPNs for connections
The initial code was gathering all of the FC ports for certain
operations.  This could lead to very long delays as the inactive ports
would run operations, and would eventually time out.

This change will only return the active ports for the VIOSes, resulting
in faster operations when dealing with storage.

Change-Id: I5c10e8e91ef6936f6c6c44644de54db3b18a41d5