Always call hdisk removal from vscsi volume driver
This change updates the vscsi volume driver to ensure that the hdisk
removal is always invoked.  The hdisk removal significantly speeds up
subsequent discoveries.  The changes to pypowervm should remove risks
from running the API.

The change is done so that an error from the hdisk job does not block
volume disconnect.  The volume will have already been disconnected from
the VM by the time the command is run.  If the hdisk remove fails for
some reason, a subsequent attach should clean it up (but can be slower
when it does so).

Change-Id: Ib823f793ec4b951111301286743c6bd243b863fe