Sanitize config ISO name
Previously, we were using the instance's name verbatim to generate the
name of its config ISO.  Since the format of the instance name is very
permissive, this can lead to rejection of the name by the PowerVM REST
API:

Value 'Linux Test_0ce757b1_thorst_config.iso' is not facet-valid with
respect to pattern '([A-Za-z0-9_\.]{1,79})' for type 'FileName.Pattern'.

This change set exploits a new pypowervm utility to sanitize the
instance name and combine it with the desired prefix and suffix to
generate a valid config ISO name.

The same is being done for an instance's boot/rescue disks and image LUs
(for SSP).

Change-Id: Id8a77e20ede3ec3339ef5e5910e05943683a2b6c