SSP: Only consider this host's VIOSes
Previously, the SSP disk driver was gleaning the list of VIOS UUIDs from
the list of Nodes in the Cluster, and optionally restricting that list
to VIOSes from a specified host.

It will almost never be useful to know about VIOSes residing on a
different host.  Any attempt to talk to those VIOSes through the PowerVM
API will fail, as we don't have a pypowervm connection to that host.

Therefore this change set simplifies the retrieval of VIOS UUIDs,
eliminating the host_uuid option and restricting the result to VIOSes on
the current host.

Change-Id: I2fe1e93b7039383889e56b809d5348ee2d5222a9