NPIV Port login-logout changes
This change includes mapping the fabric to the management
partition so that NPIV WWPNs are logged in to the FC physical
fabric. This will allow the WWPNs to be properly zoned via
cinder zone manager, as well as can be made available to the
cinder storage host.
Once the mapping is done and zones are created, later on these
mappings are removed from the management partition (logout) so
that they can be remapped to the client VM.
Different Fabric states are maintained in the instance system
metadata for each state during the entire lifecycle.
Three supported states:
'unmapped': Initial fabric state with no mapping.
'mgmt_mapped': Fabric is mapped to the management partition.
'inst_mapped': Fabric is mapped to the actual client VM instance.

Change-Id: Ic2f564ebd182024a11947e9b06ba297fe3394495