Fixes for Python 3
Nova is working on support for Python 3.4. via
https://blueprints.launchpad.net/nova/+spec/nova-python3

This changeset moves from py33 to py34 in the tox.ini file to be
inline with Nova. It also incorporates changes produced from the
'sixer' tool run with 'all'.

sixer can be found at:
https://pypi.python.org/pypi/sixer

At the present time, the py34 environment does not run
successfully because it pulls in portions of Nova that have not
yet been fully converted to Python 3.

Change-Id: If84df3017a722d0ad3546e81f97e67f1a07a763f