Add VIF plug verification to deploy
The VIF plugging operation is intended to be an atomic operation for the
VM, and it needs to complete before the VM is booted.  When using
Neutron networks, some of the operation (specifically the bridging) is
done asynchronously.

When Neutron finishes, it sends an event indicating that the operation
has completed.

This change set updates the PowerVM driver to listen for the event and
fail if it does not come in.  If Nova Networks are in place, then no
event will come back.  The code supports this scenario as well.

Change-Id: Ibb4bc519a8c07d2879d22b5206d961eb99a9b63c