Don't wait for VIF events if no creates needed
The wait for vif events from Neutron code is always invoked as it
stands.  This can cause an issue on reboots where the plug_vifs is
invoked on system init.  This is essentially part of the heal of the
system.

In this scenario, while nova does not create a VIF, it ends up waiting
for an event back from Neutron.  It never gets that event (because
Neutron has no action) and ends up timing out.

This change set rectifies this.  If there are no VIFs to create, then
the call to wait for Neutron never occurs.

Change-Id: If202bf6502a16eb12f1178f65ae01b42caed024f