Move VM state check post VIF required check
The VIF code determines if a VM is in a valid state for a plug vif, and
then determines if a plug vif is needed.  This can lead to odd scenarios
where the plug of the VIF will fail, when in fact no VIF plugs were
actually required.

This change set swaps the checks around to prevent this failure.

Change-Id: Ic759080fa2848eee179f322dd047bce3048624e8
Closes-Bug: 1466698