Remove lpar create revert method
The lpar create flow had a revert method that removed the
lpar if any exceptions were encountered.  Also the compute
manager calls the destroy operation against the instance if
any spawn exceptions are raised.  By removing the revert
method that deletes the lpar, we get a clean flow through
the destroy operation and accomplish the same result.

Change-Id: Ibdce4b1d76618678ce8b5dce473703d3ffd05c9f