Raise VirtualInterfacePlugException
The compute driver method plug_vifs() should raise
VirtualInterfacePlugException when exceptions are enountered.
This is because the compute manager calls plug_vifs() during
start up and it only handles this particular exception. If
other exceptions are raised then the compute manager fails
to start.

One particular case where this can happen is when an instance
is not found on the host.

Change-Id: Iba41727905e7732718f289ea90aa7147b763e232
Closes-Bug: #1473687