Set PowerVM uuid to OpenStack uuid
PowerVM now allows a uuid to be set at LPAR create time. The
uuid formats are not exactly the same (they differ by a single
bit) but the OpenStack uuid can be converted to a PowerVM uuid
and used at spawn time.

PowerVM uuid format always has the most significant bit set to 0.

Change-Id: I041397047f413b5142ee6b7537ad60a3c396cc80