Update vSCSI Discovery to support multiple VIOSes
PowerVM supports having multiple Virtual I/O Servers to virtualize the
I/O Adapters to the Virtual Machines.  When running a vSCSI volume
attach, the code was passing in all of the WWPNs from the host itself.
This caused the Virtual I/O Server discover hdisk to fail as it was
checking WWPNs that were not part of the VIOS.

This change set updates the query to only use the WWPNs that are
specific to a given VIOS during the discover_hdisk invocation.

Change-Id: I2a873548c6ef9fdd59f7ff6d3a15c88bb729215d
Closes-Bug: 1473989