Refactor LPAR builder code into common location
Several methods were added to the vm.py file that focused on building
the LPAR construct.  This is the translation of the Nova instance/flavor
objects into input that pypowervm can consume.  As more methods were
added to vm.py, the ones focused on building the lpar became a bit
disorganized.

This change set collapses all of the methods down into a new class
called VMBuilder.  It consolidates all of the logic required to convert
a Nova instance/flavor into the necessary pypowervm constructs.

Also added a new attribute for powervm - shared_proc_pool_name.

Change-Id: I892beb9f985baaae0c1a42829bf36144f7b3974c