Refactor to _is_booted_from_volume
The original _root_bdm_in_block_device_info method was used to determine
if the VM should be booted from a Cinder volume.  The method worked
perfectly fine, but was a bit long (causing some odd formatting issues).

This change set renames it to match its intent, _is_booted_from_volume
This makes the code a bit cleaner to read.  It is also named
consistently with the libvirt and xen methods.

Change-Id: I667729b520eb87de3ae713da5e7f98a12058460e