Unable to find PFC Ports on vSCSI volume attach
The vSCSI volume attach code needs to query for the active Virtual I/O
Servers using the xag (extended attribute) for Storage.  This will
include in the API response the detailed storage hardware information.
The code was not doing this, and as such, no Virtual I/O Servers were
candidates for the discover_hdisk.

Change-Id: Ie5dc72e1c906ac6c5e25d29f9cb40e8e7878ecd9
Closes-Bug: 1474860