Remove the UUID Cache
A recent change (set_lpar_uuid) enabled the code to tell the PowerVM API
to use the UUID that matches the Nova instance uuid upon VM creation.
This means that there is a 1 to 1 mapping between the UUIDs in OpenStack
and the PowerVM API.

Therefore, the UUID cache served little (or no) value.  This change set
removes the cache from the nova-powervm code as it is no longer needed.

Change-Id: I30072fcdf582dc10f6d193e072ddcff31668807c