Use PowerVM restart option for instance reboot
In the existing instance reboot implementation, it performs power off
then power on for the task. With PowerVM, you issue a power off
with restart option to perform reboot instead of making two
separate requests.

With this change, it will check whether instance is shutdown. If so,
it issues a power on request. Otherwise, it will issue a power off
with restart option.

Closes-Bug: #1475423

Change-Id: I1f9160ec9eae6703c2b160bde664143323c0673d