Swap to mgmt VIF being 'static link local'
In order to support the management VIF for RMC connections, the link
local address needs to be defined on the management VIF.  For legacy
versions of cloud-init as well as broad distribution support, this
change set switches to a 'static' definition of the link local address.
This applies to a broader set of cloud init revisions.

Change-Id: I2f072b697b858cb0ee5bc7189c5f593f77fede89