Add helpers to the pypowervm adapter
Helpers are a utility within the pypowervm adapter that help with edge
scenarios that can occur when interacting with the API.  This change set
adds two of the more common helpers to the ceilometer-powervm adapter
creation.

The first is the log helper.  In the event of error scenarios, this
helper will log the error.  The second is the helper for the vio busy.
In the case of a Virtual I/O Server being busy, this will retry the
requests after a short period of time.

Change-Id: I65634f394416714c452fbd9fdc48db87b9c8ed61