Add vio_retry helper to adapter
This change set adds the VIO retry helper to the default helper list on
the pypowervm adapter creation.  This allows for automated retry of
adapter calls when the API returns the 'VIOS is busy' error.

This can occur in highly concurrent scenarios, but is generally pretty
quick and non-intrusive.  This helper will make them more transparent
overall.

Change-Id: I6f3338ea7d2393083b58d208113cd099716a8310