Remove WWPN pre-mapping generation
The pypowervm API is planning to remove the method that pre-builds the
WWPNs for a given VM.  This is instead meant to be handled by letting
the mapping determine the WWPNs when the mgmt partition is mapped.

This change set updates the code such that the nova-powervm delegates
the WWPN generation to the mgmt partitions mapping.  It uses the
derive_base_npiv_map method instead.

This change set also fixes certain currency issues within the fabrics by
delegating the resp_wwpns generation to an overall for-loop.

Closes-Bug: 1478587
Change-Id: I8917779f6d9225e6460c0fa64e855c96b55fb655