Remove cpu_sink as it does not pertain to powervm
The cpu_sink as defined in the default pipeline.yaml does not apply to
PowerVM.  This is because the PowerVM compute inspector has a CPU
Utilization inspector built in.

This change provides an updated default pipeline.yaml that is applicable
for operators wishing to make use of the powervm compute inspector.

This change also removes the network sink as that was also an issue as
network metrics came back.

Change-Id: I4e7cae715fc77c5e399bd3c6d33afdc8fa02d890
Closes-Bug: 1475362