Support parallel IO next steps
This change set supports the second step of I/O parallelization that is
required for PowerVM.  It makes use of a new pypowervm function
surrounding 'Transactions'.  These transactions encapsulate retry
functions, locking, etc...

The main benefit for this is around the Virtual I/O Server changes.  The
most expensive part of the deploy path is adding/removing vSCSI and NPIV
mappings from the I/O Server to the VM.

This change allows us to batch together the I/O Server changes into a
single transaction for vDisk and NPIV operations.  A second change set
will be provided to include the config drive functions as well.

These transaction allow for us to batch together many changes to I/O
servers within a single update.  Also if multiple I/O servers are being
updated, they will each be able to run in parallel.  The multiple I/O
server changes are handled via pypowervm's transaction FeedTask.

This should allow for linear scaling regardless of the number of storage
operations that are required and independent of the number of VIOSes.

Future change sets will further enhance the amount of work delegated to
the transaction feed task.

Change-Id: I4de93f5cbdb56c337b65868c5412c69fd702e82d