Move the NPIV Mgmt Mapping into FeedTask
This change set builds on the previous speed up work.  The previous code
had the NPIV management unmapping done within a separate action from the
add to the client VM.

This change set modifies the code such that the NPIV mappings on deploy
are all done within a single transaction.

Change-Id: I5bc431b2a2391600f39b65af184cbda34c8795f5