Add vOpt media to the FeedTask
The transaction manager enables us to batch together various operations.
It is currently being used in nova-powervm to batch together the storage
mapping operations.

This change set builts on previous ones.  It adds the vOpt media to the
FeedTask (or tx_mgr).  This allows for one less independent mapping on
the spawn action.

Note that this undoes a change that has the network and the storage run
in parallel.  This is because the config drive requires input from the
network.  The overall benefit of putting the config drive in the tx_mgr
far outweighs running the network and storage in parallel.

A future change set will evaluate running all the storage and network
'creates' in parallel.

Change-Id: Ibeaa8f796372f33595650a62237206b396946e62