Remove n_spawn call the thread out the live migration
In a recent commit, the compute manager now spawns a thread for
the live migration.  This frees up the RPC call that could time
out waiting for a response of a migration that was taking a long
time.  This change set removes the spawn in the driver now that
it's done above in the compute manager.

Change-Id: Ie209e3c3b9d6d2048846f10835ccbb7e27209d10