Add proper failure on disk create revert
When a disk fails to create, the task flow revert step will be invoked.
If the failure originates from elsewhere in the task, the result will be
the disk that was created and it will be deleted properly.  But if the
error originates within the disk create step, the result is a failure.

That failure was being passed down into the delete disk step, and then
the rollback was not properly executing.  This change adds a small check
to ensure that the rollback does not get invoked if the failure
originated from that step itself.

Change-Id: I294ea9731f8325132670a70474dce1288cb73796
Closes-Bug: 1486527