Ensure LPM'd VMs don't obscure Host CPU stats
When a Virtual Machine is live migrated from one system to another, the
Host statistics are also migrated along with it.  When this occurs, the
PCM API will not have a previous sample for it.

The existing code assumed the only case that a previous sample may not
be there is if the VM had just spawned.  In which case, all of the
cycles could be attributed to just used.

However, given this LPM behavior, one can not simply add all of the CPU
cycles.  Suddenly the metric could obtain a month's (or whatever the
life of the LPM'd VM is) of data added.  The Host CPU metrics would be
too high and would be inaccurate.

This change throws out samples for VMs for which there is no previous
data.  While not ideal, it strikes a good balance for the issue.  A
subsequent sample will rectify the issue as well.  The side effect of
this is that a single sample (after a LPM occurs) may appear slightly
low on the CPU cycles.  But that is acceptable given the previous
behavior which would be wildly wrong.

Change-Id: I90c7b7c25e7b59977eb7b83823aba8b2705d69b3
Closes-Bug: 1483278