PowerVM Storage Inspector
This change set provides base support for the PowerVM compute inspector
to collect the storage devices.

Since PowerVM monitors storage metrics on the 'buses' (rather than a per
device) the metrics are reported based on the bus.  This may be a
virtual SCSI bus or a virtual Fibre Channel bus.

Change-Id: I704c7e79e825b7d5f0f8924c28d502ccbfb9330d