Account for live migrated VMs
In PowerVM, when a VM is migrated, the 'metrics come with it'.  That
means that the metrics are cumulative and you can be assured that when
it is on a new host, its metrics are still valid.

This presents some complications.  The metrics are pulled from the host.
To determine rates, we ask the host for the previous sample and the
current.  Previously we had assumed that if the previous sample was
None, it must be a new VM so all the traffic from the current is 'new'.

With live migration taken into account, we can not do that.  There may
be months of CPU, Network and Storage data in that VMs metrics.  This
really skews the metrics in undesirable ways.

This change set will skip instances (for a given inspection cycle) that
are in the current sample but not the previous.  This does lose a sample
cycle, but ensures that we maintain accurracy (which is more important).

Change-Id: Iba7f1287b3433acd8517947e01cbfa14fe93413c