Optimize the XAG query
This change set optimizes how the extended attributes are determined
within the driver.  The Storage XAG is a very expensive extended
attribute, with each query taking a couple seconds.  The MAPPING XAGs
are linear with the number of mappings on the system.

As such, it is best to determine which XAGs are needed ahead of time to
minimize the amount needed to gather.

The behavior is such that a default deploy only needs SCSI Mappings (for
vOpt and local/SSP disks).  The volume adapters vary what they need.
NPIV needs Storage and FC Mappings.  The vSCSI adapter only requires the
SCSI mappings (turns out the Storage mapping is NOT required there -
further testing showed).

Therefore a new method was introduced called _get_inst_xag.  It takes in
some information (including bdms).  If a volume is provided, it will
determine from the volume adapter what XAG attributes are required.

This should significantly speed up all deploys that do not require NPIV
volume adapter connections.  As further volume adapters are added, this
framework should be flexible to handle the new mechanisms as well.

This method also re-introduces a vSCSI optimization that was lost in a
recent refactoring.  It gathers the physical WWPNs once up front and
caches them.

Change-Id: Ia56c160950990683d70751f1d1f6ed583c6f3863