Remove local funcs in npiv.py as able
The local functions in npiv.py were buggy as they only passed in the
local variables 'for the last loop' that was run.  This led to some odd
errors.  This change set moves npiv.py away from local functions and
utilizes the standard built in functions.

This removes odd loop/local variable conditions that occurred otherwise.

Change-Id: I9621f5b9732e480398b808184d1bcadee0d9d2d1
Closes-Bug: 1488122