Scrub stale LPAR storage from spawn
Building on support in pypowervm to scrub mappings and storage elements
associated with an LPAR ID, this change set adds logic into the spawn
flow to augment the storage I/O FeedTask to do that for the LPAR ID of
the newly-created VM, in case that ID was recycled from a previously
incompletely-cleaned-up instance.

Also changed the name of the tx_mgr variable to stg_ftsk throughout, as this
will ultimately be one of multiple FeedTasks used in the various flows.

Change-Id: Id5f3201c91f3fd24a80405eb7ecfe34a641ff925