Remove parallel engine from spawn flow
Recent changes un-parallelized the spawn flow, but left the parallel
TaskFlow enigne in place.  This (probably) had the unwanted side effect
of causing the (single, linear) flow to run in a child thread, which was
causing the nova security context not to be propagated down to the
Tasks.  This change set removes the now-unnecessary use of the parallel
engine.

Recent pypowervm changes ensure that FeedTask flows will now
have the context propagated.  Future nova-powervm work that does use
parallel engines should take advantage of
pypowervm.utils.transaction.ContextThreadPoolExecutor to ensure the same
behavior.

Closes-Bug: 1487535
Change-Id: Ib3264012e060cfca16f8a2e4a0d8366fd54588b0