Remove nova specific checks on NIC operations
The IBMi operating system does not require an active 'RMC' connection
when adding or removing a virtual network interface.  There are other
variations that the base nova code does not take into account.  This
change set delegates the check if I/O can be added down to the
pypowervm wrapper instead.  This wrapper will provide a reason if the
API is not valid.

Change-Id: Iad6e046a0ec726d0d6396bb6f7f779bcf32af222