Change volume UDID key in the connection info
Previosly the volume UDID was stored in the connection information
with the key being the VIOS UUID and the volume ID.  Those data
points are not needed because the UDID does not vary with the
VIOS that it's mapped to and the volume ID is already stored
in the data section of the connection information.  This change
simplifies the storage by using a static key 'target_UDID' to
store the UDID value.

Change-Id: I2d262bb2e4b3a835214b3e27de82f294a1d721e9