Rediscover hdisk if connection info refreshed
Although the UDID for a VSCSI attached volume is saved in the bdm
during attach, it can be lost if the connection information is
'refreshed' by the compute manager.  This fix rediscovers the hdisk
during detach if we cannot look it up via the UDID. One scenario
where the connection info is refreshed is live migration.

Change-Id: I13846c494d72d57e47082d4d0fb2a286309aa0f1