Add instance save on volume attach
The volume connectors are currently saving some system metadata.  This
is specifically the NPIV connector which uses the system metadata to
store the WWPNs of the VM.

While the spawn and other lifecycle operations call instance.save, the
volume attach and detach do not.  This change provides a save invocation
on the attach.  Detach is not included as it is not critical to the bug
and may have broader impacts for live migration.

It is not put in the npiv volume connector directly as that would
guarantee a save call, when it in fact may not be required.

Change-Id: I970327fbbe1a2fc7d77a6a5951492ec24e31c11a
Closes-Bug: 1491523