Reuse existing wwwpns for live migration multi vol
When a live migration with NPIV attached volumes occurs, it is important
to flip the WWPNs around once for each fabric - but not to do it for
each volume.  This is so that the second WWPN gets logged in on the
target, but not the first (which is already logged in on the source).

This change set provides a hook in the wwpns method to derive existing.
However, this will ensure that the migration only returns the data once.
It is not a full solution however.  The end of the migration still
requires an instance save.  That will be handled in a subsequent change
set.

Change-Id: I9495bdaed1cd0ba55a95722303b552ae3c8980cd