Enable multi host live migration with NPIV Volumes
When the migration occurs the order of the WWPNs is important.  The
first WWPN is the client adapter that is logged in.  This gets
obfusicated when a live migration is occurring.  The source is primary
and logged in, and the target is secondary (also logged in).

This code adds a post migration call for the volume drivers.  It will
handle saving the updated WWPN order in the system metadata (as the
instance gets saved on the target host, post live migration).  It also
introduces a mig_vol_stor concept into post migration so that the volume
adapters can pass data from one connection to another.  This is
important for NPIV as it is concerned about fabrics rather than the
specific volumes within the fabric.

Change-Id: Ia4f3d0c35b10f6f106ba27b0fb9f408fbeb670fb