Ignore empty NPIV fabrics on disconnect
A spawn may roll back all operations within the flow, even if the
execute has not yet run against them.  This means that a disconnect
volume may occur before a connect volume if something in the flow failed
before the volume actions.

Currently, the vSCSI flow is prepared for this.  However, the NPIV path
would fail.

This change set resolves the disconnect to tolerate empty fabrics on a
disconnect.  It will detect if a fabric has not yet been connected and
log accordingly on the disconnect.

Change-Id: I78c51fa3677d78d28f1efd66a29394a24538e145
Closes-Bug: 1493940