Initial Live Migration Port Mapping Support
This change set provides support for gathering the storage mappings
for volume support during a live migration.  The storage mappings allow
the source and target servers to negotiate which fibre channel ports
should be used to connect the VMs to their volumes.

This can be very useful for NPIV based deploys.  If the client has two
fabrics, this will make sure that the appropriate fabric ports on the
target server are utilized.  This is also useful to respect the
operators port specifications in the CONF file.

Change-Id: I35efc206a2627be714c5c8b777775f485673996d